Update code based on upstream bv/fault changes
* Support for None values were removed.  Changed the functional models
to use fault.UnknownValue (X) by default. Then the test logic was
updated to expect `AnyValue` on the outputs when they are
unspecified/undefined. For example, the test expects AnyValue for
outputs of the memory during the reset sequence since the values are
undefined.

* bit_vector `as_uint()` instead of `as_int()`, since `as_int()` returns
the signed representation now (before it depended on whether the signed
flag was set on initialization).

* bit_vector `as_uint()` instead of `.unsigned_value` since the API
has changed to be consistent with the `as_int` pattern.

* Explicitly cast lists to BitVectors for tests since the BitVector
`==` operator no longer does implicit conversions.