Switch from AOS to SOA for side type
Previously the side was represented as an array-of-structs for
representing each layer (e.g. 1-bit, 16-bit). Now we switch to a
struct-of-arrays. This will make some wiring/indexing tasks easier. It
will also allow for having different numbers of tracks for each track,
e.g. only 3 1-bit tracks and 5 16-bit tracks.