Resolve dedup issue in pe core test
Magma updated how cacheing works which then caused this test to fail.
The issue is that the setup code was executed outside the test
environment, so it was not properly interacting with the magma circuit
database which is cleared between test runs.  To rememdy this issue, I
moved the pe core generator invocation inside the fixture function
instead of being in the top level module code path.