Make garnet not installable
Since garnet is a top-level repo and *not* intended as an installable
or referencable library, it should not be installed. Therefore, we have
removed setup.py and do not do `pip install -e .` in travis.

As a consequence, all requirements now are specified in requirements.txt
rather than in setup.py.