Add latch to gate the stall
We have two choices, we can either gate the enables on the registers and rely on the synthesis tools to convert these to clock gates, or we can latch the clock and do this ourselves. We implement the latch here. The latch is transparent while the clock is high, so if we get a transition on the stall right after the posedge then we don't want to create an extra early posedge with the gating logic.

This is the behavior the memory tile is expecting from the stall logic.