Add peak core/tile sequence based test
Demonstrates use of the new prototype fault sequence testing to reuse
the same input/output sequence to test the peak core and the PE_tile.

The core tester simply drives the input values and configuration and
checks the output using the core interface.

The tile tester uses a fixed routing configuration (borrowed this from
the interconnect test) to setup the switchboxes and connection boxes,
then drives the inputs onto the approprate inputs for the configuration
and checks the appropriate output for the configuration.

Future improvements:
* Support arbitrary routes and test all possible routes for tile level
  test
  * Need to abstract driver/monitor logic to select the right ports
    based on the routing configuration
* Is there an API to just get the PE_tile? Right now we generate a small
  CGRA and extra the PE tile
  * We also need a PnR API that works on the tile level (this is another
    reason we need to generate a CGRA).  Alternatively, can we just
    create all possible route configurations by construction? We still
    need to interpret these routes so we know which ports to use

Depends on https://github.com/leonardt/fault/pull/267